🚚 Rename: 계층 구조에 맞게 features/auth/ui 폴더로 이동
