+++ src/app/(public)/(auth)/layout.tsx
@@ -1,5 +1,5 @@
-import AuthBranding from '@/app/(public)/(auth)/ui/AuthBranding';
 import GuestGuard from '@/features/auth/guards/GuestGuard';
+import AuthBranding from '@/features/auth/ui/AuthBranding';
 import { cn } from '@/shared/utils/cn';
 
 /**
@@ -14,7 +14,7 @@ export default function AuthLayout({
 }>) {
   return (
     <main className="flex min-h-dvh w-full items-center justify-center xl:justify-between">
-      <div className={cn('mx-auto w-full max-w-187 px-6 py-16 md:px-13.5')}>
+      <div className={cn('mx-auto w-full max-w-187 px-6 py-13 md:px-13.5')}>
         <GuestGuard>{children}</GuestGuard>
       </div>
       <div
