+++ src/features/auth/ui/SignupClient.tsx
@@ -3,11 +3,11 @@ import { zodResolver } from '@hookform/resolvers/zod';
 import { useRouter } from 'next/navigation';
 import { useCallback } from 'react';
 import { useForm } from 'react-hook-form';
-import AuthHeadline from '@/app/(public)/(auth)/ui/AuthHeadline';
 import { signupUser } from '@/features/auth/apis/auth';
 import { AUTH_API_MESSAGE } from '@/features/auth/constants/authMessage';
 import AuthFooter from '@/features/auth/ui/AuthFooter';
 import AuthForm from '@/features/auth/ui/AuthForm';
+import AuthHeadline from '@/features/auth/ui/AuthHeadline';
 import { SignupFormValues, signupSchema } from '@/features/auth/utils/schema';
 import { ApiError } from '@/shared/apis/apiError';
 import Button from '@/shared/ui/button/Button';
